move admin-related assertions to its own module
