+++ ca/django_ca/tests/test_admin_acme.py
@@ -27,7 +27,7 @@ from django_ca.models import (
     AcmeOrder,
     CertificateAuthority,
 )
-from django_ca.tests.base.assertions import assert_changelist_response
+from django_ca.tests.admin.assertions import assert_changelist_response
 from django_ca.tests.base.mixins import StandardAdminViewTestCaseMixin
 from django_ca.tests.base.typehints import DjangoCAModelTypeVar
 from django_ca.tests.base.utils import override_tmpcadir
